Able to get all pe ratios for all stocks in S&P500
The endpoint `http://127.0.0.1:8000/stocks/pe_ratios/` will query for
pe ratios for all tickers in db.

I tried yfinance (https://pypi.org/project/yfinance/), but it kinda
sucks. It is also slow and fails for a lot of tickers. I switched to
scraping pe ratios myself on yahoo finance using BeautifulSoup.

It is really slow though. I plan on saving the pe ratios in the db so I
only have to query for them once a day and just load from the db other
times.